Adjusting data blocks for LOAD segments
LOAD segments can only load data when the offset of that data and its
destination address are equal modulo page size.  Section content is now
aligned according to that constraint (I'm using the page offsets of the
original file to be able to recreate original address space layout), but
the segment table still needs to take the new addresses into account.