Introduce layoutDescription.phdr_vaddr
The layoutDescription.phdr_vaddr member stores the virtual address of
the PHDR table in the new file. This replaces the computation done from
the new file address and the original base loading address, which relied
on a continuous mapping of the file from file header to PHDR table.

Additional small changes:
* Alignment of PHDR table is now parametrized via PHDR32ALIGN &
  PHDR64ALIGN.
* Flags of section ranges are now ORed on aggregation to segment ranges