Fixed alignment requirements for ranges in sections
Removed old requirement that a range to keep has to be aligned with the
section alignment. Instead it is now required that the range starts and
ends at addresses that are aligned with entry size of its surrounding
section.

Also changed fill byte to an interrupt generating machine code (on x86).