shrinkelf: improve readability, fix deprecated addConstr warning
In order to make the construction of the SMT constraints more
readable, let's compute the constant values once and give them
an intuitive name. We can also add a few more constraints to
the solving process which state some (more or less obvious)
requirements to a valid solution.

Furthermore, gurobi suggests using addLConstr instead of
addConstr which also speeds up the solving process so let's
follow that.