shrinkelf: fix shift calculations if first fragment has offset > 0 from beginning of section
If the first fragment inside a section already starts at an
offset to the original beginning of the section (for example,
if the first function in the target library was deleted), the
ILP strategy would generate non-PAGESIZE-d values for the
shift members of the contained fragments which ultimately
leads to a failing elf_update() call because the section
will become unaligned.

Fix this up by moving the start of the section back to the
original start in this case.