shrinkelf: fix logic for PHDR attachement at the end of a segment
With enough functions removed from an ELF file, the first shift
of a segment can occur very early (after .dynsym, for example).

In this case, we might need to place the program header in a
gap that does not actually exist yet. While there already was
some logic in place, it did not work correctly and missed the
case that we might need to attach it to the previous fragment.

This commit fixes the situation for glibc 2.28 and clang, where
the program header needs to be placed quite early in the file and
all following sections need to be moved up by two pages.