shrinkelf: speed up insertRange and insertTuple
As the incoming lists should already be sorted, we can speed
up insertion by doing bisection instead of plain iteration
over all values.

I think the checks should be equivalent for sorted lists as
overlaps can only happen if (a) the end of the inserted
fragment is larger than the start of the following fragment
or (b) the end of the previous fragment goes beyond the
start of the inserted fragment.