shrinkelf: fix section start if first fragment starts at an offset
While we already implemented this (partially) for the ILP
case, the SMT process still had the problem that a section
would be misaligned if the first fragment inside that section
already had an offset (i.e., the first bytes of a section
were deleted). This led to a 'shifted' start of the section
and thus calculated wrong shifts for all fragments and the
section as a whole.

To fix this, check the first fragment and shift the
calculated beginning of the section back by the offset of the
first fragment.

In the ILP case, we only considered the case of a fixed first
fragment so far but I think we also need to do this for other
fragments placed in the first slot - we still need to check
that the calculations work correctly in that case.