shrinkelf: obey section alignment while reordering
While reordering fragments inside a section, we could
actually move the first fragment to a place where the
original alignment of the section (which we carry over from
the input file) would no longer be obeyed by the beginning
of the section. This leads to failing elf_update() calls
as the section headers are checked before writing them to
the file.

Fix this by moving the lowest address possible for the
following section to the next aligned boundary before
running the optimization process.