shrinkelf.py: fix handling of empty target sections
If the ranges to keep skipped entire sections from files, we
did not correctly handle missing entries in the computed
section ranges and failed to compute an output file.

To handle this, we check if the keep ranges provided leave out
entire sections and generate a 1-byte pseudo entry for these
sections. To accomplish this more easily, we refactored some
parts of the parsing process and mapping the ranges to
sections in the input file.