shrinkelf.py: fix misaligned fragments special case
If a section starts with a non-fixed first fragment, the solution
could place the first fragment into the same page as the end of the
previous section if the respective offsets allow merging the pages
together. However, the previously introduced logic for fixing up the
start of the section if the first fragment starts at an offset with
respect to the original section start could then move the logical
beginning of the section into the previous section.

This should not happen as it will lead to inconsistent shift values
and misaligned addresses and offsets in the section header table.

Fix it up by checking if the 'fixed' start of the section reaches
into the already occupied space - if it does, move the beginning of
the current section forward by a whole page to allow starting at
the next free boundary.