Add tests
This adds a test binary file for shrinking. The libtest.nasm file
is the source code for the shipped libtest.so file, which has
ranges of functions which can be compacted by reordering them.

The test cases run all variants (z3, gurobi, brute-force and
shifting), and we will add a comparison/checks for optimality
in a following commit.