shrinkelf.py: fix size calculations for brute-force case
The logic for a fixed first fragment in the section was not
the same as for the other cases, so align it with the SMT
and ILP logic.

Additionally, when calculating the size of a particular
permutation, we need to take the potential additional
offset of a given fragment from the beginning of the current
section into account.